Errors made, OrdersService complete

+++ Project2Api.Contracts/Order/OrderRequest.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 namespace Project2Api.Contracts.Order
 {
     public record OrderRequest(
-        List<string> Items,
-        int Price
+        List<string> Items, 
+        float Price
     );
 }
